Cleaning up the gruntfile, before adding a grunt task generator
*	Moved grunt tasks/options to ./lib/tasks
*	Now you can easily work on a grunt task without the mess which is Gruntfile.js
*	Gruntfile.js is now about 50loc
*	Preparation work for a `sails g task` command
