+++ bin/boilerplates/lib/tasks/build.task.js
@@ -1,0 +1,10 @@
+module.exports = function(grunt){
+  // Build the assets into a web accessible folder.
+  // (handy for phone gap apps, chrome extensions, etc.)
+  grunt.registerTask('build', [
+    'compileAssets',
+    'linkAssets',
+    'clean:build',
+    'copy:build'
+  ]);
+};
