+++ bin/boilerplates/lib/tasks/compileAssets.task.js
@@ -1,0 +1,9 @@
+module.exports = function(grunt){
+  grunt.registerTask('compileAssets', [
+    'clean:dev',
+    'jst:dev',
+    'less:dev',
+    'copy:dev',
+    'coffee:dev'
+  ]);
+};
