+++ bin/boilerplates/lib/tasks/default.task.js
@@ -1,0 +1,7 @@
+module.exports = function(grunt) {
+  grunt.registerTask('default', [
+    'compileAssets',
+    'linkAssets',
+    'watch'
+  ]);
+};
