+++ bin/boilerplates/lib/tasks/fromSails.task.js
@@ -1,0 +1,14 @@
+module.exports = function(grunt) {
+  // Get path to core grunt dependencies from Sails
+  var depsPath = grunt.option('gdsrc') || 'node_modules/sails/node_modules';
+  grunt.loadTasks(depsPath + '/grunt-contrib-clean/tasks');
+  grunt.loadTasks(depsPath + '/grunt-contrib-copy/tasks');
+  grunt.loadTasks(depsPath + '/grunt-contrib-concat/tasks');
+  grunt.loadTasks(depsPath + '/grunt-sails-linker/tasks');
+  grunt.loadTasks(depsPath + '/grunt-contrib-jst/tasks');
+  grunt.loadTasks(depsPath + '/grunt-contrib-watch/tasks');
+  grunt.loadTasks(depsPath + '/grunt-contrib-uglify/tasks');
+  grunt.loadTasks(depsPath + '/grunt-contrib-cssmin/tasks');
+  grunt.loadTasks(depsPath + '/grunt-contrib-less/tasks');
+  grunt.loadTasks(depsPath + '/grunt-contrib-coffee/tasks');
+};
