+++ bin/boilerplates/lib/tasks/linkAssets.task.js
@@ -1,0 +1,11 @@
+module.exports = function(grunt){
+  grunt.registerTask('linkAssets', [
+    // Update link/script/template references in `assets` index.html
+    'sails-linker:devJs',
+    'sails-linker:devStyles',
+    'sails-linker:devTpl',
+    'sails-linker:devJsJADE',
+    'sails-linker:devStylesJADE',
+    'sails-linker:devTplJADE'
+  ]);
+};
