+++ bin/boilerplates/lib/tasks/options/clean.js
@@ -1,0 +1,4 @@
+module.exports = {
+  dev: ['.tmp/public/**'],
+  build: ['www']
+};
