+++ bin/boilerplates/lib/tasks/options/concat.js
@@ -1,0 +1,11 @@
+var commonConfig = require('../common/common.js');
+module.exports = {
+  js: {
+    src: commonConfig.jsFilesToInject,
+    dest: '.tmp/public/concat/production.js'
+  },
+  css: {
+    src: commonConfig.cssFilesToInject,
+    dest: '.tmp/public/concat/production.css'
+  }
+};
