+++ bin/boilerplates/lib/tasks/options/cssmin.js
@@ -1,0 +1,6 @@
+module.exports = {
+  dist: {
+    src: ['.tmp/public/concat/production.css'],
+    dest: '.tmp/public/min/production.css'
+  }
+};
