+++ bin/boilerplates/lib/tasks/options/sails-linker/devJsJADE.js
@@ -1,0 +1,13 @@
+var commonConfig = require('../../common/common.js');
+
+module.exports = {
+  options: {
+    startTag: '// SCRIPTS',
+    endTag: '// SCRIPTS END',
+    fileTmpl: 'script(type="text/javascript", src="%s")',
+    appRoot: '.tmp/public'
+  },
+  files: {
+    'views/**/*.jade': commonConfig.jsFilesToInject
+  }
+};
