+++ bin/boilerplates/lib/tasks/options/sails-linker/prodJsJADE.js
@@ -1,0 +1,11 @@
+module.exports = {
+  options: {
+    startTag: '// SCRIPTS',
+    endTag: '// SCRIPTS END',
+    fileTmpl: 'script(type="text/javascript", src="%s")',
+    appRoot: '.tmp/public'
+  },
+  files: {
+    'views/**/*.jade': ['.tmp/public/min/production.js']
+  }
+};
