+++ bin/boilerplates/lib/tasks/options/uglify.js
@@ -1,0 +1,6 @@
+module.exports = {
+  dist: {
+    src: ['.tmp/public/concat/production.js'],
+    dest: '.tmp/public/min/production.js'
+  }
+};
