+++ bin/boilerplates/lib/tasks/options/watch.js
@@ -1,0 +1,15 @@
+module.exports = {
+  api: {
+
+    // API files to watch:
+    files: ['api/**/*']
+  },
+  assets: {
+
+    // Assets to watch:
+    files: ['assets/**/*'],
+
+    // When assets are changed:
+    tasks: ['compileAssets', 'linkAssets']
+  }
+};
