+++ bin/boilerplates/lib/tasks/prod.task.js
@@ -1,0 +1,19 @@
+module.exports = function(grunt){
+  // When sails is lifted in production
+  grunt.registerTask('prod', [
+    'clean:dev',
+    'jst:dev',
+    'less:dev',
+    'copy:dev',
+    'coffee:dev',
+    'concat',
+    'uglify',
+    'cssmin',
+    'sails-linker:prodJs',
+    'sails-linker:prodStyles',
+    'sails-linker:devTpl',
+    'sails-linker:prodJsJADE',
+    'sails-linker:prodStylesJADE',
+    'sails-linker:devTplJADE'
+  ]);
+};
