+++ test/cli/integration/new.test.js
@@ -223,7 +223,38 @@ function checkGeneratedFiles(appName, templateLang) {
 		'config/locales/en.json',
 		'config/locales/es.json',
 		'config/locales/fr.json',
-		'config/locales/de.json'
+		'config/locales/de.json',
+    'lib',
+    'lib/tasks',
+    'lib/tasks/build.task.js',
+    'lib/tasks/compileAssets.task.js',
+    'lib/tasks/default.task.js',
+    'lib/tasks/fromSails.task.js',
+    'lib/tasks/linkAssets.task.js',
+    'lib/tasks/prod.task.js',
+    'lib/tasks/options/sails-linker',
+    'lib/tasks/options/sails-linker/devJs.js',
+    'lib/tasks/options/sails-linker/devJsJADE.js',
+    'lib/tasks/options/sails-linker/devStyles.js',
+    'lib/tasks/options/sails-linker/devStylesJADE.js',
+    'lib/tasks/options/sails-linker/devTpl.js',
+    'lib/tasks/options/sails-linker/devTplJADE.js',
+    'lib/tasks/options/sails-linker/prodJs.js',
+    'lib/tasks/options/sails-linker/prodJsJADE.js',
+    'lib/tasks/options/sails-linker/prodStyles.js',
+    'lib/tasks/options/sails-linker/prodStylesJADE.js',
+    'lib/tasks/options',
+    'lib/tasks/options/clean.js',
+    'lib/tasks/options/coffee.js',
+    'lib/tasks/options/concat.js',
+    'lib/tasks/options/copy.js',
+    'lib/tasks/options/cssmin.js',
+    'lib/tasks/options/jst.js',
+    'lib/tasks/options/less.js',
+    'lib/tasks/options/uglify.js',
+    'lib/tasks/options/watch.js',
+    'lib/tasks/common',
+    'lib/tasks/common/common.js'
 	];
 
 	// Add template files of the specified language
